Auth
init

Update login-form.tsx

update

update

update

Create actions.tsx

Update schema.prisma

Update user-nav.tsx

Update schema.prisma

Update schema.prisma

Update schema.prisma

routes updated ( still require checkup )

useing `setSession()` instead of `getUser()`

Update route.ts

Update route.ts

removed components

Update actions.ts

login with github

Update actions.ts

Update OAuthButtons.tsx

Update page.tsx

Update middleware.ts

Update page.tsx

Update page.tsx

Update page.tsx

Update page.tsx

+++ app/api/forms/[id]/delete/route.ts
@@ -1,5 +1,6 @@
 import { NextResponse } from 'next/server'
 import { PrismaClient } from '@prisma/client'
+import { createClient } from '@/utils/supabase/server'
 
 const prisma = new PrismaClient()
 
@@ -7,8 +8,30 @@ export async function DELETE(
   request: Request,
   { params }: { params: { id: string } }
 ) {
+  const supabase = await createClient()
+  const {
+    data: { session },
+  } = await supabase.auth.getSession()
+
+  if (!session?.user.id) {
+    return NextResponse.json({ error: 'Unauthorized' }, { status: 401 })
+  }
+
   try {
     const id = params.id
+
+    // Validate the form exists
+    const existingForm = await prisma.forms.findUnique({
+      where: { id },
+    })
+
+    if (!existingForm) {
+      return NextResponse.json({ error: 'Form not found' }, { status: 404 })
+    }
+
+    if (existingForm.userId !== session.user.id) {
+      return NextResponse.json({ error: 'Unauthorized' }, { status: 403 })
+    }
 
     const deletedForm = await prisma.forms.delete({
       where: { id },
